Moved Message class into the default package (terrible idea but getting tired to write code correctly)
Implemented A* search to find home when computer is found for student
class. (This only works half the time).

There are two major bugs that occur randomly. First: Agents can't get
the message to return home when being told by other agents. Second:
There is a bug in the A* search heuristic that will cause it to infinite
loop sometimes.