Cleaned up some debugging code.
Added AgentRunnerHelper class to help initialize agents. Just need to
modify the settings in the method and it'll create all the agents we
want.

Added Student2.java which is a test Student agent that should appear
"smarter" compared to Student. Unfortunately it infinite loops sometimes
and then it looks like an idiot.

Added a "done" condition so the GUI knows when to stop animating and can
load new rooms.

Added several more rooms to have fun with.